Add files via upload
In this study I compared a serial algorithm with MPI and Agapia algorithms to compute the air quality index and standard deviation of a dataset.

I used the Air Pollution in Seoul dataset from Kaggle (https://www.kaggle.com/bappekim/air-pollution-in-seoul). With the python script "CreateFiles" I extracted only one element from the dataset and created csv files of different size to compare the efficiency of each algorithm. Because of uploading constraints the csv I included is very small. You can use the python script to create bigger files.